requirements.txt
discord
python-decouple
pandas
